requirements/requirements.txt
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.0
